enable admin dashboard: registrations - modularise components
